smart-news/requirement.txt
python-jsonrpc
pymongo
pika
requests
redis
lxml
goose-extractor
python-dateutil
sklearn
pandas
tensorflow
watchdog

